+++ PartB/Ex1/Lang.h
@@ -9,7 +9,7 @@ class Lang{
     string analysisText;
 
     public:
-        Lang(int k, int alpha, string refText, string analysisText);
+        Lang(int k, float alpha, string refText, string analysisText);
         map<string, map<char,int>> getRefTextModel();
         map<string, map<char,int>> getAnalysisTextModel();
         double calculateBits(map<string, map<char,int>> refTxtModel, map<string, map<char,int>> analysisTxtModel);
