+++ PartB/Ex2/Findlang.h
@@ -6,12 +6,12 @@ using namespace std;
 
 class Findlang{
     int k;
-    int alpha;
+    float alpha;
     vector<string> refTexts;
     string analysisText;
 
     public:
-        Findlang(int k, int alpha, vector<string> refTexts, string analysisText);
+        Findlang(int k, float alpha, vector<string> refTexts, string analysisText);
         vector<map<string, map<char,int>>> getRefTextsModels();
         map<string, map<char,int>> getAnalysisTextModel();
         string guessLanguage(vector<map<string, map<char,int>>> refTextsModels, map<string, map<char,int>> analysisTextModel);
